+++ txn.go
@@ -375,8 +375,8 @@ func (txn *Txn) modify(e *Entry) error {
 		// keep things safe and allow badger move prefix and a timestamp suffix, let's
 		// cut it down to 65000, instead of using 65536.
 		return exceedsSize("Key", maxKeySize, e.Key)
-	case int64(len(e.Value)) > txn.db.opt.ValueLogFileSize:
-		return exceedsSize("Value", txn.db.opt.ValueLogFileSize, e.Value)
+	case uint32(len(e.Value)) > txn.db.opt.ValueLogFileSize:
+		return exceedsSize("Value", int64(txn.db.opt.ValueLogFileSize), e.Value)
 	case txn.db.opt.InMemory && len(e.Value) > txn.db.opt.ValueThreshold:
 		return exceedsSize("Value", int64(txn.db.opt.ValueThreshold), e.Value)
 	}
